Extract common double prim code
